refactor: rename organization packages
Renames `groupproject` and `usergroupmember` packages.
Similar to resource names.
