+++ internal/plugin/service/organization/groupproject/organization_group_project.go
@@ -1,4 +1,4 @@
-package project
+package groupproject
 
 import (
 	"context"
