+++ internal/plugin/service/organization/groupproject/organization_group_project_test.go
@@ -1,4 +1,4 @@
-package project_test
+package groupproject_test
 
 import (
 	"fmt"
