+++ internal/plugin/service/organization/usergroupmember/organization_user_group_member.go
@@ -1,4 +1,4 @@
-package usergroup
+package usergroupmember
 
 import (
 	"context"
