+++ internal/plugin/service/organization/usergroupmember/organization_user_group_member_test.go
@@ -1,4 +1,4 @@
-package usergroup_test
+package usergroupmember_test
 
 import (
 	"fmt"
